renames 'mode.type.ts' to 'view.type.ts'
